tests and new properties class
