--- src/integration-test/java/uk/ac/ebi/tsi/portal/api/deployment/WithUserDetailsSecurityContextFactory.java
@@ -1,37 +1,0 @@
-package uk.ac.ebi.tsi.portal.api.deployment;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContext;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.test.context.support.WithSecurityContextFactory;
-import org.springframework.security.test.context.support.WithUserDetails;
-import org.springframework.stereotype.Component;
-import org.springframework.util.Assert;
-
-public class WithUserDetailsSecurityContextFactory implements WithSecurityContextFactory<WithUserDetails> {
-
-    private CustomUserDetailsService customUserDetailsService;
-
-    @Autowired
-    public WithUserDetailsSecurityContextFactory(CustomUserDetailsService userDetailsService) {
-        this.customUserDetailsService = userDetailsService;
-    }
-
-    public SecurityContext createSecurityContext(WithUserDetails withUser) {
-        String username = withUser.value();
-        Assert.hasLength(username, "value() must be non-empty String");
-        UserDetails principal = customUserDetailsService.loadUserByUsername(username);
-        Authentication authentication = new UsernamePasswordAuthenticationToken(principal, principal.getPassword(), principal.getAuthorities());
-        SecurityContext context = SecurityContextHolder.createEmptyContext();
-        context.setAuthentication(authentication);
-        return context;
-    }
-
-}
