+++ src/integration-test/resources/integrationTest.properties
@@ -44,4 +44,4 @@ ajayPassword=${AAP_AJAY_PASSWORD}
 pantherUserName=${AAP_PANTHER_USERNAME}
 pantherPassword=${AAP_PANTHER_PASSWORD}
 
-ecp.roles.deployments.view=ROLE_self.VIEW_ROLE_TEST
+ecp.aap.domains.deployments.view=ROLE_self.VIEW_ROLE_TEST
