+++ src/main/resources/application.properties
@@ -105,4 +105,4 @@ ecp.scriptUser=@user.name@
 #Spring boot config enable for BASE URL from request
 server.use-forward-headers=true
 
-ecp.roles.deployments.view=ROLE_self.AUTH_PORTAL
+ecp.aap.domains.deployments.view=ROLE_self.AUTH_PORTAL
